mcp_knowledge/requirements.txt
# MCP Knowledge Server Dependencies

# Ragflow API
aiohttp>=3.9.0

# MCP Server
mcp>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
pydantic>=2.5.0
python-dotenv